refactor: Move fxsend/fxrec test-utility main() bodies into cmd/
fxsend.go and fxrec.go only ever existed to be invoked as standalone
FX.25 test binaries, so move their logic directly into
cmd/samoyed-fxsend/main.go and cmd/samoyed-fxrec/main.go rather than
routing through exported *Main() functions in the direwolf package.

The fx25_init/fx25_send_frame/fx25_rec_bit/fx25_test_count internals
they relied on stay unexported, so add src/fx25_wrapper.go exposing
thin wrappers for cmd/ to call.
