--- src/fxsend.go
@@ -1,25 +1,0 @@
-//nolint:gochecknoglobals
-package direwolf
-
-import (
-	"fmt"
-)
-
-var preload = []byte{
-	'T' << 1, 'E' << 1, 'S' << 1, 'T' << 1, ' ' << 1, ' ' << 1, 0x60,
-	'W' << 1, 'B' << 1, '2' << 1, 'O' << 1, 'S' << 1, 'Z' << 1, 0x63,
-	0x03, 0xf0,
-	'F', 'o', 'o', '?', 'B', 'a', 'r', '?', //  '?' causes bit stuffing
-}
-
-func FxsendMain() {
-	fmt.Println("fxsend - FX.25 unit test.")
-	fmt.Println("This generates 11 files named fx01.dat, fx02.dat, ..., fx0b.dat")
-	fmt.Println("Run fxrec as second part of test.")
-
-	fx25_init(3)
-
-	for i := 100 + CTAG_MIN; i <= 100+CTAG_MAX; i++ {
-		fx25_send_frame(0, preload, i, true)
-	}
-}
